Refactor directory inode responsibilities
Separate real backend directories from synthetic virtual directories by trimming DirINode down to the real-tree behavior and keeping the virtual-node implementation in VirtualDirINode. Shared inode cache, locking, and path helpers stay in baseDirNode.

Also move synthetic inode generation into the shared base helper, update root inode construction to use the shared constructor, and remove the obsolete virtual-kind assertion from the collision test.
